backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
pydantic[email]==2.10.5
supabase==2.11.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.28.1
groq==0.15.0

# ML deps (shared with src/)
numpy>=1.26.4
pandas>=2.2.2
pyarrow>=15.0.2
scikit-learn>=1.4.2
xgboost>=2.0.3
shap>=0.45.1
joblib>=1.3.2
